# api/openapi.yaml
openapi: 3.1.0
info:
  title: Portwing Agent API
  version: 0.1.0
  description: |
    Portwing is a lightweight remote Docker agent for the Drydock container
    monitoring platform. This specification covers the **Standard Mode** HTTP API
    only (Edge Mode communicates over a WebSocket multiplexed protocol documented
    separately in SPEC.md §3).

    ## Authentication

    All protected endpoints require the agent token, supplied via one of:

    - `Authorization: Bearer <token>` (standard, checked first)
    - `X-Portwing-Token` header
    - `X-Dd-Agent-Secret` header (Drydock agent secret header)

    Token comparison uses `crypto/subtle.ConstantTimeCompare` (timing-safe).
    If no token is configured, authentication is disabled and all requests pass
    through unauthenticated.

    **Rate limiting:** 10 failed attempts per IP per rolling minute. Exceeding
    the threshold returns HTTP 429 until the window expires.

    ## Docker Engine API Proxy

    Any path not matched by a named route is transparently proxied to the local
    Docker daemon (Unix socket). This includes every endpoint in the
    [Docker Engine API](https://docs.docker.com/engine/api/) — streaming
    endpoints (`/logs`, `/attach`, `/exec/*/start`, `/events`, `/build`,
    `/images/create`, `/images/push`), interactive exec hijacking
    (`Upgrade: tcp`), and binary responses. Paths are forwarded verbatim; the
    agent negotiates the API version on startup (fallback: `v1.44`).

...

  contact:
...
servers:
...
components:
  securitySchemes:
    bearerToken:
      type: http
      scheme: bearer
      description: >
        Agent token via Authorization: Bearer. Also accepted in the
        `X-Portwing-Token` header, or Drydock's `X-Dd-Agent-Secret` header.
    portwingToken:
      type: apiKey
      in: header
      name: X-Portwing-Token
      description: Agent token via the X-Portwing-Token header.
    ed25519Signature:
      type: apiKey
      in: header
      name: X-Portwing-Signature
      description: |
        Ed25519 per-request signature (base64url, no padding). When this header
        is present the agent verifies the request via Ed25519 instead of the
        shared token.

        Four headers are required together:

        | Header | Format | Example |
        |--------|--------|---------|
        | `X-Portwing-Key-ID` | hex(SHA-256(raw 32-byte pubkey)[:8]) | `4a3f2b1c9d8e7f6a` |
        | `X-Portwing-Timestamp` | Unix epoch, seconds, decimal | `1749600000` |
        | `X-Portwing-Nonce` | 32 random hex chars (128-bit) | `a1b2c3d4e5f6...` |
        | `X-Portwing-Signature` | base64url (no padding) Ed25519 signature | `d2hh...` |

        The signed canonical string is:
        ```
        METHOD\nPATH\nbody-sha256-hex\nunix-timestamp\nnonce
        ```
        For an empty body, `body-sha256-hex` is the full SHA-256 of the empty
        string: `e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855`.

        On failure, the response includes `X-Portwing-Reason` with one of:
        `timestamp-skew`, `replay`, `unknown-key`, `invalid-signature`.
  schemas:
...
    Container:
      type: object
      description: Drydock container model with update metadata.
      properties:
        id:
          type: string
          example: "abc123def456"
        name:
          type: string
          example: "myapp-web-1"
        displayName:
          type: string
          example: "My App Web"
        displayIcon:
          type: string
          example: "docker"
        status:
          type: string
          example: "running"
        watcher:
          type: string
          example: "docker"
        agent:
          type: string
        image:
          $ref: '#/components/schemas/ContainerImage'
        result:
          $ref: '#/components/schemas/ContainerResult'
        error:
          $ref: '#/components/schemas/ContainerError'
        updateAvailable:
          type: boolean
        updateKind:
          type: string
          example: "tag"
        includeTags:
          type: string
        excludeTags:
          type: string
        transformTags:
          type: string
        labels:
          type: object
          additionalProperties:
            type: string
        details:
          $ref: '#/components/schemas/RuntimeDetails'
    ContainerImage:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        tag:
          type: string
    ContainerResult:
      type: object
      properties:
        tag:
          type: string
        created:
          type: string
          format: date-time
    ContainerError:
      type: object
      properties:
        message:
          type: string
    RuntimeDetails:
      type: object
      description: Runtime container details (ports, mounts, environment, etc.).
      additionalProperties: true
    WatcherComponent:
      type: object
      description: Drydock watcher component descriptor.
      additionalProperties: true
    TriggerComponent:
      type: object
      description: Drydock trigger component descriptor.
      additionalProperties: true
    NotImplementedResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: "not implemented in v1.0"
        message:
          type: string
          example: "registry checking is performed by the Drydock controller"
...
paths:
...
  /api/events:
    get:
      operationId: getApiEvents
      summary: SSE container event stream
      description: >
        Server-Sent Events stream of container lifecycle events. Immediately
        emits a `dd:ack` event with agent metadata and current container
        counts, then streams `dd:container-added`, `dd:container-updated`,
        and `dd:container-removed` events as Docker reports them.

        The connection is held open until the client disconnects. Each event
        is a JSON object on a `data:` line followed by a blank line (standard
        SSE format).

        **Event types:**

        | Type | Description |
        |------|-------------|
        | `dd:ack` | Initial handshake; agent metadata + container counts |
        | `dd:container-added` | New container discovered |
        | `dd:container-updated` | Container state or metadata changed |
        | `dd:container-removed` | Container removed |

        **Example stream:**
        ```
        data: {"type":"dd:ack","data":{"version":"0.1.0","os":"linux","arch":"amd64","cpus":4,"memoryGb":0,"uptimeSeconds":3600,"lastSeen":"2024-07-23T10:00:00Z","containers":{"total":5,"running":4,"stopped":1},"images":3}}

        data: {"type":"dd:container-added","data":{...Container...}}

        data: {"type":"dd:container-removed","data":{"id":"abc123","name":"old-container"}}
        ```
      tags: [Drydock Compat]
      security:
        - bearerToken: []
        - portwingToken: []
      responses:
        "200":
          description: SSE stream opened.
          content:
            text/event-stream:
              schema:
                type: string
                description: Newline-delimited SSE frames.
        "401":
          description: Missing or invalid token.
        "429":
          description: Rate limit exceeded.
        "500":
          description: Streaming not supported by server.
  /api/containers:
    get:
      operationId: getApiContainers
      summary: Container inventory
      description: >
        Returns the full container inventory as parsed from Docker with
        `dd.*` label metadata applied. This is the same data set sent via
        the `dd:container_sync` WebSocket message in Edge mode.
      tags: [Drydock Compat]
      security:
        - bearerToken: []
        - portwingToken: []
      responses:
        "200":
          description: Array of containers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Container'
        "401":
          description: Missing or invalid token.
        "429":
          description: Rate limit exceeded.
  /api/containers/{id}/logs:
    get:
      operationId: getApiContainerLogs
      summary: Container logs
      description: >
        Returns demuxed container logs (Docker's 8-byte stream multiplexing
        header is stripped). Supports optional tailing, time filtering, and
        live following.
      tags: [Drydock Compat]
      security:
        - bearerToken: []
        - portwingToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Container ID or name.
          example: "abc123def456"
        - name: tail
          in: query
          schema:
            type: integer
            minimum: 1
          description: Number of lines from the end to return.
          example: 100
        - name: since
          in: query
          schema:
            type: string
          description: >
            Show logs since this timestamp (RFC3339) or relative duration
            (e.g., `1h`). Passed directly to Docker API.
          example: "2024-07-23T10:00:00Z"
        - name: until
          in: query
          schema:
            type: string
          description: Show logs before this timestamp or duration.
        - name: follow
          in: query
          schema:
            type: string
            enum: ["0", "1", "true", "false"]
          description: >
            Stream logs in real time. Response uses chunked transfer
            encoding when `follow=1` or `follow=true`.
      responses:
        "200":
          description: Log output (plain text, demuxed).
          content:
            text/plain:
              schema:
                type: string
        "400":
          description: Invalid `tail` value.
        "401":
          description: Missing or invalid token.
        "429":
          description: Rate limit exceeded.
        "500":
          description: Error retrieving logs from Docker.
  /api/containers/{id}:
    delete:
      operationId: deleteApiContainer
      summary: Remove a container
      description: Removes a container (force=true). Returns 204 on success.
      tags: [Drydock Compat]
      security:
        - bearerToken: []
        - portwingToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Container ID or name.
      responses:
        "204":
          description: Container removed.
        "401":
          description: Missing or invalid token.
        "429":
          description: Rate limit exceeded.
        "500":
          description: Error removing container.
  /api/watchers:
    get:
      operationId: getApiWatchers
      summary: Watcher component descriptors
      description: >
        Returns the list of watcher component descriptors registered with
        this agent. In v1.0, Portwing reports a single `docker` watcher;
        registry checks are delegated to the Drydock controller.
      tags: [Drydock Compat]
      security:
        - bearerToken: []
        - portwingToken: []
      responses:
        "200":
          description: Array of watcher components.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WatcherComponent'
        "401":
          description: Missing or invalid token.
        "429":
          description: Rate limit exceeded.
  /api/triggers:
    get:
      operationId: getApiTriggers
      summary: Trigger component descriptors
      description: Returns the list of trigger component descriptors.
      tags: [Drydock Compat]
      security:
        - bearerToken: []
        - portwingToken: []
      responses:
        "200":
          description: Array of trigger components.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TriggerComponent'
        "401":
          description: Missing or invalid token.
        "429":
          description: Rate limit exceeded.
  /api/watchers/{type}/{name}:
    post:
      operationId: postApiWatcherPoll
      summary: Trigger watcher poll
      description: >
        Not implemented in v1.0. Registry checking is performed by the
        Drydock controller. Returns HTTP 501.
      tags: [Drydock Compat]
      security:
        - bearerToken: []
        - portwingToken: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
          example: "docker"
        - name: name
          in: path
          required: true
          schema:
            type: string
          example: "my-watcher"
      responses:
        "501":
          description: Not implemented.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotImplementedResponse'
        "401":
          description: Missing or invalid token.
        "429":
          description: Rate limit exceeded.
  /api/watchers/{type}/{name}/container/{id}:
    post:
      operationId: postApiWatcherContainer
      summary: Check single container
      description: >
        Not implemented in v1.0. Registry checking is performed by the
        Drydock controller. Returns HTTP 501.
      tags: [Drydock Compat]
      security:
        - bearerToken: []
        - portwingToken: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "501":
          description: Not implemented.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotImplementedResponse'
        "401":
          description: Missing or invalid token.
        "429":
          description: Rate limit exceeded.
  /api/triggers/{type}/{name}:
    post:
      operationId: postApiTriggerExec
      summary: Execute trigger
      description: >
        Not implemented in v1.0. Returns HTTP 501.
      tags: [Drydock Compat]
      security:
        - bearerToken: []
        - portwingToken: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "501":
          description: Not implemented.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotImplementedResponse'
        "401":
          description: Missing or invalid token.
        "429":
          description: Rate limit exceeded.
  /api/triggers/{type}/{name}/batch:
    post:
      operationId: postApiTriggerBatch
      summary: Execute batch trigger
      description: >
        Not implemented in v1.0. Returns HTTP 501.
      tags: [Drydock Compat]
      security:
        - bearerToken: []
        - portwingToken: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        "501":
          description: Not implemented.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotImplementedResponse'
        "401":
          description: Missing or invalid token.
        "429":
          description: Rate limit exceeded.
...
  /api/v1/containers:
    get:
      operationId: getGenericContainers
      summary: Container inventory (generic adapter)
      description: >
        Returns the cached container inventory built from the Docker daemon.
        Available when `ADAPTER=generic`. The response shape is the same
        `Container` model used by the Drydock-compatible endpoints.
      tags: [Generic]
      security:
        - bearerToken: []
        - portwingToken: []
      responses:
        "200":
          description: Array of containers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Container'
        "401":
          description: Missing or invalid token.
        "429":
          description: Rate limit exceeded.
  /api/v1/containers/{id}/logs:
    get:
      operationId: getGenericContainerLogs
      summary: Container logs (generic adapter)
      description: >
        Returns demuxed container logs (Docker's 8-byte stream multiplexing
        header is stripped). Supports optional tailing, time filtering, and
        live following. Available when `ADAPTER=generic`.
      tags: [Generic]
      security:
        - bearerToken: []
        - portwingToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Container ID or name.
          example: "abc123def456"
        - name: tail
          in: query
          schema:
            type: integer
            minimum: 1
          description: Number of lines from the end to return.
          example: 100
        - name: since
          in: query
          schema:
            type: string
          description: Show logs since this timestamp (RFC3339) or relative duration.
          example: "2026-01-01T00:00:00Z"
        - name: until
          in: query
          schema:
            type: string
          description: Show logs before this timestamp or duration.
        - name: follow
          in: query
          schema:
            type: string
            enum: ["0", "1", "true", "false"]
          description: Stream logs in real time with chunked transfer encoding.
      responses:
        "200":
          description: Log output (plain text, demuxed).
          content:
            text/plain:
              schema:
                type: string
        "400":
          description: Invalid `tail` value.
        "401":
          description: Missing or invalid token.
        "429":
          description: Rate limit exceeded.
        "500":
          description: Error retrieving logs from Docker.
  /api/v1/events:
    get:
      operationId: getGenericEvents
      summary: Docker lifecycle event stream (generic adapter)
      description: >
        Server-Sent Events stream of Docker container lifecycle events in a
        stable generic envelope. Available when `ADAPTER=generic`.

        Each event is a JSON object on a `data:` line followed by a blank line
        (standard SSE format). A comment heartbeat (`": heartbeat"`) is sent
        every 30 seconds to keep the connection alive. The stream reconnects
        internally with exponential backoff if Docker becomes unavailable.

        Client disconnect cleanly terminates the stream.

        **Event actions** (from `allowedActions` in `internal/docker/events.go`):
        `create`, `start`, `stop`, `die`, `kill`, `restart`, `pause`,
        `unpause`, `destroy`, `rename`, `update`, `oom`, `health_status`.

        **Example stream:**
        ```
        data: {"ts":"2026-06-11T10:00:00Z","type":"container","action":"start","containerId":"abc123","name":"my-app","image":"nginx:latest"}

        : heartbeat

        data: {"ts":"2026-06-11T10:00:30Z","type":"container","action":"die","containerId":"abc123","name":"my-app","image":"nginx:latest","labels":{"app":"web"}}
        ```
      tags: [Generic]
      security:
        - bearerToken: []
        - portwingToken: []
      responses:
        "200":
          description: SSE stream opened.
          content:
            text/event-stream:
              schema:
                type: string
                description: >
                  Newline-delimited SSE frames. Each `data:` frame contains a
                  `GenericEvent` JSON object.
        "401":
          description: Missing or invalid token.
        "429":
          description: Rate limit exceeded.
        "500":
          description: Streaming not supported by server.
...